feat: :sparkles: Employees CRUD

+++ src/services/employee.service.ts
@@ -1,55 +1,70 @@
-import { Employee } from "../entities/employee"
+import { Employee, EmployeeRequest } from "../entities/employee"
+import { CompanyRepository } from "../repositories/company.repository"
 import { EmployeeRepository } from "../repositories/employee.repository"
-import { InMemoryEmployeeRepository } from "../repositories/in-memory/in-memory.employee.repository"
-import { DuplicatedEmployeeError } from "./errors/DuplicatedEmployeeError"
-import { EmployeeNotFoundError } from "./errors/EmployeeNotFoundError"
-import { Service } from "./service"
+import { PrismaCompanyRepository } from "../repositories/prisma/prisma.company.repository"
+import { PrismaEmployeeRepository } from "../repositories/prisma/prisma.employee.repository"
+import { BadRequestError, ConflictError, NotFoundError } from "../shared/api-errors"
 
 
-export class EmployeeService implements Service<Employee, string> {
-  employeeRepository: EmployeeRepository
+export class EmployeeService {
+  private employeeRepository: EmployeeRepository
+  private companyRepository: CompanyRepository
   constructor(
-    repository?: EmployeeRepository
+    repositoryEmployee?: EmployeeRepository,
+    repositoryCompany?: CompanyRepository
   ) {
-    this.employeeRepository = repository ?? new InMemoryEmployeeRepository()
+    this.employeeRepository = repositoryEmployee ?? new PrismaEmployeeRepository()
+    this.companyRepository = repositoryCompany ?? new PrismaCompanyRepository()
   }
 
   async findAll(): Promise<Employee[]> {
     return this.employeeRepository.findAll()
   }
 
-  async find(id: string): Promise<Employee | null> {
-    return this.employeeRepository.find(id)
+  async find(id: string): Promise<Employee> {
+    const result = await this.employeeRepository.find(id)
+    if (!result) {
+      throw new NotFoundError('Employee not found')
+    }
+    return result
   }
 
-  async create(employee: Employee): Promise<Employee> {
-    const result = await this.employeeRepository.findAll()
+  async create(employee: EmployeeRequest): Promise<Employee> {
+    const employees = await this.employeeRepository.findAll()
 
-    result.forEach(repEmployee => {
+    employees.forEach(repEmployee => {
       if (repEmployee.id === employee.id) {
-        throw new DuplicatedEmployeeError()
+        throw new ConflictError('Invalid create employee with duplicated id')
       }
     })
 
-    const newEmployee = await this.employeeRepository.create(employee)
-    return this.employeeRepository.save(newEmployee)
+    const result = await this.companyRepository.find(employee.company_id)
+    if (!result) {
+      throw new BadRequestError('Invalid company id')
+    }
+
+    return this.employeeRepository.save(employee)
   }
 
-  async update(id: string, company: Partial<Employee>): Promise<Employee> {
-    try {
-      const result = await this.employeeRepository.update(id, company)
-      return result
-    } catch (error) {
-      throw new EmployeeNotFoundError()
+  async update(id: string, employee: EmployeeRequest): Promise<Employee> {
+    if (!await this.employeeRepository.find(id)) {
+      return this.create(employee)
     }
+
+    const companies = await this.employeeRepository.findAll()
+    companies.forEach(repEmployee => {
+      if (repEmployee.id === employee.id) {
+        throw new ConflictError('Invalid create employee with duplicated id')
+      }
+    })
+
+    return this.employeeRepository.update(id, employee)
   }
 
   async delete(id: string): Promise<Employee> {
-    try {
-      const result = await this.employeeRepository.delete(id)
-      return result
-    } catch (error) {
-      throw new EmployeeNotFoundError()
+    if (!await this.employeeRepository.find(id)) {
+      throw new NotFoundError('Employee not found')
     }
+    return this.employeeRepository.delete(id)
   }
 }
